% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasta_to_table.R
\name{fasta_to_table}
\alias{fasta_to_table}
\title{Convert fasta to table.}
\usage{
fasta_to_table(.fasta, .col_name = X1)
}
\arguments{
\item{.fasta}{fasta format files.}

\item{.col_name}{name of column.}
}
\description{
converting fasta format file to table that have two rows including  OTU and sequence.
}
\examples{

fasta <-
   data.frame(X1 = c(">otu_1", "ATGCGG", ">otu_2", "GCCGA"))

fasta_to_table(.fasta = fasta, .col_name = X1)
}
