% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_unifrac.R
\name{get_unifrac}
\alias{get_unifrac}
\title{Get UniFrac distance.}
\usage{
get_unifrac(.com, .tree, .alpha = 0.5, .data_type)
}
\arguments{
\item{.com}{community data matrix.}

\item{.tree}{phylogenetic tree (phylo object).}

\item{.alpha}{controlling the weight on abundant.}

\item{.data_type}{"W", "UW", "VAW". see GUniFrac::GUniFrac help.}
}
\description{
Get UniFrac distance.
}
