% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_seq.R
\name{load_seq}
\alias{load_seq}
\title{Load sequence and sample infomation from NCBI by using E-utilities.}
\usage{
load_seq(.id, .path = "download.xml", .fasta = T, .term = F)
}
\arguments{
\item{.id}{character, Genbank accession ID.}

\item{.path}{saved xml file name.}

\item{.fasta}{only sequence.}

\item{.term}{logical, you understand E-Utilities term ?}
}
\description{
laod fungal sequence and sample data from NCBI using accession ID.
You should keep to the regulations of E-utilities.
}
