% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_index.R
\name{make_index}
\alias{make_index}
\title{Making Index primer fasta file.}
\usage{
make_index(.path, .skip = 18, .id, .index_1, .index_2)
}
\arguments{
\item{.path}{file path.}

\item{.skip}{integer, number of skipping row, default 18}

\item{.id}{id column}

\item{.index_1}{index1 column}

\item{.index_2}{index2 column}
}
\description{
This function make index_1.fasta and index_2.fasta from SampleSheet.csv.
}
