% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_to_tibble.R
\name{matrix_to_tibble}
\alias{matrix_to_tibble}
\title{Convert matrix to tibble.}
\usage{
matrix_to_tibble(.mat, .row = FALSE, .rowname = NULL)
}
\arguments{
\item{.mat}{matrix}

\item{.row}{logical, convert rowname to column?}

\item{.rowname}{character, rowname.}
}
\description{
Convert matrix to tibble.
}
\examples{
mat <-
   matrix(data = rnorm(100),
          nrow = 10,
          ncol = 10)

rownames(mat) <- sprintf("com_\%d", 1:10)

# convert to tibble.
matrix_to_tibble(.mat = mat, .row = FALSE)
# move rownames to column.
matrix_to_tibble(.mat = mat, .row = TRUE, .rowname = "com")

}
