% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_blastxml.R
\name{parse_blastxml}
\alias{parse_blastxml}
\title{Parse blast output xml}
\usage{
parse_blastxml(.xml, .col = NULL, .multi = FALSE)
}
\arguments{
\item{.xml}{blast output xml.}

\item{.col}{colnames. (not inplemented)}

\item{.multi}{logical, (not inplemented)}
}
\description{
Parse blast output xml
}
