% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_gbxml.R
\name{parse_gbxml}
\alias{parse_gbxml}
\title{Parse GenBank output xml}
\usage{
parse_gbxml(.xml, .gbseq = c("all", "non_seq", "manual"))
}
\arguments{
\item{.xml}{blast output xml.}

\item{.gbseq}{GBseq list.}
}
\description{
Parse GenBank output xml
}
