% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_number_to_group.R
\name{row_number_to_group}
\alias{row_number_to_group}
\title{Parse row number to grouping id.}
\usage{
row_number_to_group(.df, .col_of_row_number, .group)
}
\arguments{
\item{.df}{tibble, data.frame}

\item{.col_of_row_number}{name of row number column.}

\item{.group}{name of group column.}
}
\description{
This function make index_1.fasta and index_2.fasta from SampleSheet.csv.
}
\examples{

df <-
   data.frame(x = seq(1, 10, 1),
              y = c(1, 2, 3, 1, 2, 1, 2, 3, 4, 5))

row_number_to_group(df, .col_of_row_number = "y", .group = "group")

}
