% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_seed.R
\name{search_seed}
\alias{search_seed}
\title{search best seed in nMDS}
\usage{
search_seed(n = 100, .data, k = 2, trymax = 1000, maxit = 1000, ...)
}
\arguments{
\item{n}{number of seed}

\item{.data}{data}

\item{k}{dimention}

\item{trymax}{trymax}

\item{maxit}{maxit}

\item{...}{see metaMDS}
}
\description{
search best seed in nMDS
}
