% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_to_fasta.R
\name{table_to_fasta}
\alias{table_to_fasta}
\title{Convert table to fasta.}
\usage{
table_to_fasta(.data, .otu, .seq)
}
\arguments{
\item{.data}{data.frame}

\item{.otu}{otu column name.}

\item{.seq}{sequence column name.}
}
\description{
Convert table to fasta.
}
\examples{

library(tibble)

table <-
   tibble(otu = c("otu_1", "otu_2"),
          seq = c("ATGGG", "GCCGTA"))

table_to_fasta(.data = table, .otu = otu, .seq = seq)

}
