% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tibble_to_dist.R
\name{tibble_to_dist}
\alias{tibble_to_dist}
\title{Convert tibble to dist.}
\usage{
tibble_to_dist(.dist_tbl, .matrix = F)
}
\arguments{
\item{.dist_tbl}{tibble, data.frame}

\item{.matrix}{output}
}
\description{
This function is inverse operation of broom::tidy() against dist object.
}
