% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_dist.R
\name{beta_dist}
\alias{beta_dist}
\title{Calculate beta diversity.}
\usage{
beta_dist(.method, .comm, .tree, .weighted)
}
\arguments{
\item{.method}{dissimilarity index.}

\item{.comm}{community data matix.}

\item{.tree}{an object of class "phylo".}

\item{.weighted}{logical, presence/absence or abundance.}
}
\description{
Calculate beta diversity.
}
